Initial commit
Apple Music Ultra is a custom Music client based on the web beta (https://beta.music.apple.com). As of this build, it contains 16 custom NSVisualEffect themes that can be applied via the in-app menu bar (see Menu > Themes > [ThemeList]).

I will be injecting custom CSS and JS to customize the app, as well as some native integration.

In a brief: a better Apple Music.