Cleaned & reorganized project hierarchy
Removed macOS directory (created it in the first place for possible iOS port). if there were to ever be an iOS version, it would most likely be a native client as the AM Web Player Beta doesn't allow for mobile devices.

However, I would like for the primary focus to be on the macOS app, as there are more issues with the native Music app. Furthermore, people seem to be far more interested in the macOS version. This is probably due to the lack of custom Music clients for macOS and plenty of clients for iOS.