Added ThemeController, Update Check and NSAlert
Started on the ThemeController... still not sure what the most efficient way of going about this is. Any help or suggestions would be highly apprecoated.

Similarily added an update check that compares build number to a server-side TXT file (will go all-in on update checks in future builds, once development is heavier).

CHANGES
• Added ThemeController with starter variables
• Fixed title issue (loaded pages now have an accurate page title value)
• Removed unused commented code
• Added update checker (compares CFBundleVersion to TXT file on server)

BUGS
• Update check against server-side TXT file isn't accurate due to cache issues

TXT cache issue: https://stackoverflow.com/questions/60583944/loading-reading-fresh-string-from-a-txt-file-on-an-https-server-without-cache