Project Restructure, Tidy Code & Documentation
Restructured project hierarchy for better organizational practices. Furthermore, added easy-to-use helper functions for tidier code that's easier to read. Finally, added project MARKers, function documentation and removed unused code.

CHANGES
• Restructured project with groups
• Spread out code to proper groups
• Added helper functions
• Added documentation to functions & code
• Changed function names for convention
• Cleaned and formatted code for easier reading
• Categorized ViewController with MARKers
• Rearranged ViewController code for logic
• Prepare code for new structure changes to Themes
• Move from Defaults to live variable types
• More reliable saving to Defaults (rework in-progress)

HELPERS
• cleanURL: Cleans and reformats Optional URL string
• cssToString: Extracts Optional contents of a CSS file, cleans it and then reformats it as a String for use in WKWebView

STRUCTS
• Music: ViewController -> Client
• User: ViewController -> Client
• Style: ViewController -> Style

GROUPS
• Login_Controllers -> Login
• Item, TestData -> Preferences
• PrefViewController -> Preferences
• ThemeController, Settings -> Themes
• Apple_NowPlayable -> NowPlaying

BUGS
Saving Defaults: Currently I'm saving multiple variables to the UserDefaults, constantly. Furthermore, I *was* previously using them to set active themes and make changes on the fly. Details below...

I knew this wasn't the way to go about things; however, for the sake of heavy development and the new theme structure I'm working on, I was looking for a temporary means of passing things through UserDefaults as they're set to become active.

With the new structure – when passing variables through Defaults – I'd like to send a single Theme object, as opposed to each theme property (as it is now, when saving). Due to the nature of UserDefaults, this will make it much easier to track down what's getting saved to memory and what's not. Furthermore, as I have in recent commits, I'd like to handle active theme-ing with live variables – only passing the Theme object through to Defaults when the application becomes inactive or is being exited.

This will allow for reliable saving of the user's current session settings, and restoring those session settings upon the next launch.

There are, of course, more bugs that need to be covered. However, this is what I'm focusing on now.